+++ libs/dot-components/src/lib/ctm/workitem/WorkItemTooltip.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import { useStylesWithRootClass } from '../../components/useStylesWithRootClass';
+import { CommonProps } from '../../components/CommonProps';
+import { DotIcon } from '../../components';
+import { rootClassName, StyledTooltipTitle } from './WorkItemTooltip.styles';
+import { Tooltip } from '@material-ui/core';
+import {
+  ProgressionBoardThemeProvider,
+  ThemeOptions,
+} from '../progression-board/ThemeProvider';
+
+export interface WorkItemTooltipProps extends CommonProps {
+  isSplit: boolean;
+  value_goal: string;
+  title: string;
+  external_key: string;
+  child: JSX.Element;
+  theme?: ThemeOptions;
+}
+
+export const WorkItemTooltip = ({
+  className,
+  'data-testid': dataTestId,
+  isSplit,
+  value_goal,
+  title,
+  external_key,
+  child,
+  theme,
+}: WorkItemTooltipProps) => {
+  const rootClasses = useStylesWithRootClass(className);
+  const truncate = (str: string) => {
+    return str.length > 20 ? str.substring(0, 17) + '...' : str;
+  };
+
+  const populateTooltip = (isSplit: boolean) => {
+    const truncated_title = truncate(title);
+    return (
+      <ProgressionBoardThemeProvider theme={theme}>
+        <StyledTooltipTitle className={rootClassName}>
+          <div className="tooltip-header">
+            <DotIcon
+              className={value_goal}
+              iconId={isSplit ? 'circle-half-full' : 'circle'}
+              fontSize="small"
+            />
+            {external_key}
+          </div>
+          {truncated_title}
+        </StyledTooltipTitle>
+      </ProgressionBoardThemeProvider>
+    );
+  };
+
+  return (
+    <Tooltip
+      data-testid={dataTestId}
+      className={rootClasses}
+      title={populateTooltip(isSplit)}
+    >
+      {child}
+    </Tooltip>
+  );
+};
+
+export default WorkItemTooltip;